refactoring daos
