+++ README.md
@@ -1,7 +1,7 @@
 # 🎵 SongLib
 
 **SongLib** is a modern, cross-platform church hymn library and songbook manager.
-With **14 built-in songbooks** and powerful search, it lets you keep all your favorite hymns right at your fingertips — even offline.
+With **10+ songbooks**, **10000+ songs** and powerful search, it lets you keep all your favorite hymns right at your fingertips — even offline.
 
 Available for **Android, iOS, Windows, Linux, and macOS**.
